Fix for issue reported to mace-opensaml-users.
Add overloaded methods to Decrypter which optionally makes the cached DOM Element of the decrypted and unmarshalled
XMLObject the root of a new DOM Document.  This allows DOM-level ID resolution (and therefore standalone signature
verification) to work correcly on the XMLObject.  See extensive docs in the Decrypter class.
Add unit test for same.